asl-api/requirements.txt
imutils~=0.5.4
scikit-learn~=0.24.2
tqdm~=4.61.0
pandas~=1.2.4
Django~=3.2.3
django-cors-headers
django-environ
djangorestframework~=3.12.4
dj_database_url
gunicorn
whitenoise
opencv-python~=4.5.2.54
albumentations~=0.5.2
joblib~=1.0.1
numpy~=1.20.3
matplotlib~=3.4.2
nltk~=3.6.2
asgiref~=3.3.4
pillow~=8.2.0
